fix: install Lambda dependencies for prod (#133)
The source for this module must be local since it requires that the
`aws/app/lambda` dependencies are installed before running
`terraform plan/apply`.  If a remote Terraform module is used,
this step can't happen.

A future fix to this would be using Docker images to create
the Lambda functions so that no dependency install step
is required.