feat: add Open ID Connect roles
Please note this is a first step to get OIDC enabled in your project,
this will create the required roles for planning and applying terraform
changes.

After this change is put in place we can switch the Github Workflows to
authenticate using those new roles and then finally delete the terraform
users and secrets.

I've also only updated the staging workflows as I want to make sure that
staging works before we apply changes in production