chore: use GitHub app token with release-please
Update the Release Please workflow to use a GitHub app token
generated using the SRE Read/Write bot which has the required
`Content read/write` and `PR read/write` permissions.

The temporary token generated will allow for the invoke of dependent
workflows and be revoked once the workflow is finished.