fix: remove use of `always()` in the TF apply jobs
The use of `always()` makes it impossible to cancel the workflow.
This change means that the workflow will still run as expected
if jobs are skipped, but allow for cancel actions as well.

Additionally, the check on the `github.ref` has been removed from
the Staging workflow as this workflow only triggers on push
to the `develop` branch.