feat: simplify production release reverts
Update the Production Terraform workflows so that a release
can more easily be reverted.  Now all that will be required to
revert a failed release is to revert the Release Please PR.

The way this works is as follows:
1. When a PR that updates `version.txt` is opened, the TF plan
workflow will check if it's a Release Please branch.  If yes, the
branch is used to perform the TF plan.  Otherwise the tag
specified in `version.txt` is used for the TF plan operation.

2. On merge of a PR with a `version.txt` change, the TF apply
workflow will wait for the git tag it specifies to exist.  For a
normal release flow, this will be once the `release_generator`
workflow finished.  For a revert PR, the tag will already exist.

Additionally, the code checkout step has been removed from
the `update-lambda-function-image` job as it is not needed.