feat: add API ECS service to the Forms cluster (#734)
Add an API ECS service to the Forms cluster. This also includes a new ECR to hold the 
API's Docker image.

As with the IdP, a feature flag has been added to control the creation of API infrastructure.
This will give us fine-grained control over when resources are created in each environment.