fix: downgrade local AWS provider
Downgrade the local AWS provider as the latest version is
calling AWS APIs that have not been implemented in
localstack yet.

Remove the remaining .terraform.lock.hcl files.  Our provider
and module versions are all pinned to versions so this should
not have an impact on our TF action repeatability.