fix: downgrade local AWS provider (#743)
Downgrade the local AWS provider as the latest version is calling AWS APIs that have
not been implemented in localstack yet:

```
InternalFailure: the describe_addresses_attribute action has not been implemented
```

Remove the remaining .terraform.lock.hcl files. Our provider and module versions are all
pinned to versions so this should not have an impact on our TF action repeatability.