feat: run Nagware everyday
Update the Nagware function schedule so that it run everyday.
This is being done so that the Redis cache of overdue responses
is kept more up-to-date.

The logic of the Nagware has also been updated so that emails are
still only sent Tuesday/Thursday and Slack messages posted on
Sunday.