chore: downgrade to Zitadel v2.61.1 (#833)
There appears to be a regression with Zitadel v2.62.0 where the configuration 
is not being read properly. For now we can go to v2.61.1 and see if a fix is
released for the problem.

As part of this change, the config string values are being quoted as
recommended by the Zitadel docs.