test: revert IdP ECS/RDS resources
Revert the IdP resources for now since the performance tests
are complete.  Going forward as demand grows, these will likely
need to be increased.

Note that the max ACUs have been increased from their original
values as the database is bottlenecked under load at 2 ACUs.

The rationale for dropping the ECS resources below minimum recommneded
values is that the current IdP demand is very low and the response
times are acceptable for that demand.