test: revert IdP ECS/RDS resources (#858)
Revert the IdP resources for now since the IdP performance tests are complete. 
As demand grows, these will likely need to be increased.

Note that the RDS max ACU has been increased from its original value as the 
database is bottlenecked under load at 2 ACUs.

The rationale for dropping the ECS resources below minimum recommended 
values (4 vCPU, 8 GB memory) is that the current IdP demand is very low and 
the response times are acceptable for that demand.