feat: add SSM parameters for the load tests (#872)
Add AWS SSM parameters needed by the load tests to run. This includes changes to the 
Lambda so that it can read these parameters and load them into the execution environment.

The reason these are not being passed in directly as Lambda environment variables is 
because these are stored as plain text and can be leaked by AWS CLI commands.