feat: update the processed and historical data paths (#941)
Update the data paths in preparation for transferring this data to the Platform data lake.

This is being done since the S3 replication rules that will be used to perform the data 
transfer cannot alter the replicated file's destination prefix. As such, we're updating 
the files to have the desired prefix to fit into the data lake's organization structure.